pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "peloton-cli"
version = "0.1.0"
description = "Minimal Peloton CLI with pluggable API backends"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "peloton",
    "python-dotenv>=1.0.1",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.scripts]
peloton-cli = "peloton_cli.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.3",
    "black>=24.0.0",
    "pre-commit>=3.6.0",
]

[tool.uv.sources]
peloton = { git = "https://github.com/geudrik/peloton-client-library.git" }

[tool.hatch.build.targets.wheel]
packages = ["src/peloton_cli"]
